+++ tests/ecmascript/test-bi-proxy-target-proxy-2.js
@@ -1,0 +1,11 @@
+/*===
+123
+===*/
+
+var curr = { foo: 123 };
+for (var i = 0; i < 1e6; i++) {
+    curr = new Proxy(curr, {});
+}
+
+var finalProxy = curr;
+print(finalProxy.foo);